podman: Update alpine-latest-clang
Building new a new Alpine fails on the docker image generation with the error:

 ```
 ERROR: unable to select packages:
   libunwind-dev-1.8.1-r0:
     breaks: llvm-libunwind-19.1.4-r0[!libunwind-dev]
     satisfies: world[libunwind-dev]
 ```

Use the package `llvm-libunwind-dev` instead.

The `build.sh` is updated to allow read/write access to the `/source` folder. An additional check is made if building with no commands in non-interactive mode (it can still be used to build the container).

+++ qnx/scripts/build.sh
@@ -119,7 +119,11 @@ if [ ${ROOT} -ne 0 ]; then
   SOURCE="rw"
 else
   ROOTOPT="--userns=keep-id"
-  SOURCE="ro"
+  if [ ${INTERACTIVE} -ne 0 ]; then
+    SOURCE="rw"
+  else
+    SOURCE="ro"
+  fi
 fi
 
 BUILDDIR=""
@@ -136,7 +140,11 @@ fi
 if [ ${INTERACTIVE} -ne 0 ]; then
   podman run -it --rm $ROOTOPT -v $PWD/$BASEDIR:/source:${SOURCE} -v "$PWD/$BASEDIR/qnx/build/$BUILDDIR":/build:rw --tmpfs /tmp "coderesearch:$PODVERSION"
 else
-  echo "Non-Interactive"
-  echo "$OTHERARGS"
-  podman run -t --rm $ROOTOPT -v $PWD/$BASEDIR:/source:${SOURCE} -v "$PWD/$BASEDIR/qnx/build/$BUILDDIR":/build:rw --tmpfs /tmp "coderesearch:$PODVERSION" sh -c "$OTHERARGS"
+  if [ "x$OTHERARGS" = "x" ]; then
+    echo "No command, doing nothing..."
+  else
+    echo "Non-Interactive"
+    echo "$OTHERARGS"
+    podman run -t --rm $ROOTOPT -v $PWD/$BASEDIR:/source:${SOURCE} -v "$PWD/$BASEDIR/qnx/build/$BUILDDIR":/build:rw --tmpfs /tmp "coderesearch:$PODVERSION" sh -c "$OTHERARGS"
+  fi
 fi
